+++ packages/framework/resources/views/pages/documentation-search.blade.php
@@ -1,12 +1,9 @@
 @extends('hyde::layouts.docs')
 @section('content')
     <h1>Search the documentation site</h1>
-    <style>#search-menu-button, .edit-page-link {
-            display: none !important;
-        }
+    <style>#search-menu-button, .edit-page-link { display: none !important; }</style>
 
-        #search-results {
-            max-height: unset !important;
-        }</style>
-    <x-hyde::docs.search-input class="max-w-xs border-b-4 border-indigo-400"/>
+    <div class="not-prose">
+        <x-hyde::docs.hyde-search class="max-w-sm" :modal="false" />
+    </div>
 @endsection